insights_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
# ML
scikit-learn>=1.5.0
vaderSentiment>=3.3.2
bertopic>=0.16.0
sentence-transformers>=3.0.0
statsmodels>=0.14.0
numpy>=1.26.0
pandas>=2.2.0
